Restrict entropy/data to the predefined cases in the BIP39 document
The implementation is still very rough and I've left in the code paths
handling the general case of arbitrary length data.
I'm still not convinced either way if the algorithm is possible to
extend in such a way: since it there might not be a unique way to
recover the input data length uniquely from the number of encoded bits.